backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
httpx==0.27.2
stripe==10.12.0
redis==5.1.1
celery==5.4.0
sendgrid==6.11.0
boto3==1.35.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
cryptography==43.0.1
slowapi==0.1.9
email-validator==2.1.0
pydantic[email]
